+++ backend/src/main/java/org/tarjetamish/user/service/UserService.java
@@ -3,6 +3,7 @@ package org.tarjetamish.user.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.tarjetamish.user.dto.UserDTO;
+import org.tarjetamish.user.mapper.IUserConverter;
 import org.tarjetamish.user.model.User;
 import org.tarjetamish.user.repository.UserRepository;
 
@@ -14,27 +15,25 @@ import java.util.Optional;
 public class UserService {
 
     private final UserRepository userRepository;
+    private final IUserConverter userConverter;
 
     public List<UserDTO> list() {
         return userRepository.findAll().stream()
-                .map(this::convertToDTO)
+                .map(userConverter::toUserDTO)
                 .toList();
     }
 
     public Optional<UserDTO> findById(Long id) {
-        return Optional.of(userRepository.findById(id).map(this::convertToDTO).orElse(null));
+        return Optional.of(userRepository.findById(id).map(userConverter::toUserDTO).orElse(null));
     }
 
     public UserDTO save(UserDTO userDTO) {
-        User userEntity = new User(userDTO.id(), userDTO.rut(), userDTO.name(), userDTO.email(), userDTO.pin());
-        return convertToDTO(userRepository.save(userEntity));
+        return userConverter.toUserDTO(
+                userRepository.save(userConverter.toUserEntity(userDTO))
+        );
     }
 
     public void deleteUser(Long id) {
         userRepository.deleteById(id);
-    }
-
-    private UserDTO convertToDTO(User user) {
-        return new UserDTO(user.getId(), user.getRut(), user.getName(), user.getEmail(), user.getPin());
     }
 }
